refactor(core): centralize file type constants
