+++ .changeset/file-types-module.md
@@ -1,0 +1,6 @@
+---
+'@lapidist/design-lint': patch
+---
+
+export file type mappings and glob patterns from dedicated module
+
