+++ docs/architecture.md
@@ -36,7 +36,7 @@ Supplies raw text to the linter. The Node adapter [`FileSource`](https://github.
 Convert documents into ASTs. CSS uses PostCSS, JavaScript and TypeScript rely on the TypeScript compiler, and Vue/Svelte files compile before analysis. See [`src/core/framework-parsers`](https://github.com/bylapidist/design-lint/tree/main/src/core/framework-parsers).
 
 ### File type mapping
-File extensions are normalised via [`FILE_TYPE_MAP`](https://github.com/bylapidist/design-lint/blob/main/src/core/linter.ts). This mapping tells the linter which parser to apply. To support a new extension, add it to `FILE_TYPE_MAP` and implement or reference the appropriate parser.
+File extensions and default file search patterns live in [`file-types.ts`](https://github.com/bylapidist/design-lint/blob/main/src/core/file-types.ts). The exported `FILE_TYPE_MAP` tells the linter which parser to apply, while `defaultPatterns` lists the globs used to discover files. To support a new extension, add it to `FILE_TYPE_MAP` (and `defaultPatterns` if needed) and implement or reference the appropriate parser.
 
 ### Rule engine
 Registers rules and coordinates execution. See [`src/core/linter.ts`](https://github.com/bylapidist/design-lint/blob/main/src/core/linter.ts) and [`src/core/rule-registry.ts`](https://github.com/bylapidist/design-lint/blob/main/src/core/rule-registry.ts).
